+++ docs/cargo_auto_lib/auto_encrypt_decrypt_with_ssh_mod/sidebar-items.js
@@ -1,0 +1,1 @@
+window.SIDEBAR_ITEMS = {"fn":["decrypt_with_ssh_from_json","encrypt_with_ssh_interactive_save_json"],"struct":["EncryptedString","SecretBytes","SecretString"]};