+++ src/auto_delete_old_js_snippets_mod.rs
@@ -3,11 +3,12 @@
 //! deletes old js snippets when working with wasm-pack
 
 //region: use statements
-use crate::public_api_mod::{RED, RESET};
+use crate::public_api_mod::{RED, RESET, YELLOW};
 use filetime::FileTime;
 use std::env;
 use std::fs;
 use std::path::PathBuf;
+
 //endregion
 
 /// deletes old js snippets when working with wasm-pack  
@@ -42,7 +43,7 @@ pub fn auto_delete_old_js_snippets() {
                     // if second_mtime > first_mtime {
                     std::cmp::Ordering::Greater => {
                         let first_folder = opt_first_folder.unwrap();
-                        println!("    delete first: {:?}", first_folder);
+                        println!("    {YELLOW}delete first: {:?}{RESET}", first_folder);
                         std::fs::remove_dir_all(first_folder).unwrap();
 
                         opt_first_folder = Some(second_folder.clone());
@@ -50,7 +51,7 @@ pub fn auto_delete_old_js_snippets() {
                     }
                     //  } else if first_mtime > second_mtime {
                     std::cmp::Ordering::Less => {
-                        println!("    delete second: {:?}", second_folder);
+                        println!("    {YELLOW}delete second: {:?}{RESET}", second_folder);
                         std::fs::remove_dir_all(second_folder).unwrap();
                     }
                     // else